Move metrics helpers
