+++ frontend/src/metabase/metrics-viewer/utils/__tests__/test-helpers.ts
@@ -1,8 +1,11 @@
 import { createMockMetadata } from "__support__/metadata";
 import * as Lib from "metabase-lib";
 import * as LibMetric from "metabase-lib/metric";
-import Metadata from "metabase-lib/v1/metadata/Metadata";
-import type { NormalizedMetric } from "metabase-types/api";
+import type Metadata from "metabase-lib/v1/metadata/Metadata";
+import {
+  createMetricMetadata,
+  setupDefinition,
+} from "metabase-lib/v1/metric/test-helpers";
 import { createMockMeasure } from "metabase-types/api/mocks/measure";
 import {
   createMockMetricDimension,
@@ -109,26 +112,7 @@ export const GEO_DIM_IDX = {
   DATE_TIME: 5,
 };
 
-export function createMetricMetadata(metrics: NormalizedMetric[]): Metadata {
-  const metadata = new Metadata();
-  for (const metric of metrics) {
-    const { collection: _normalizedCollectionId, ...rest } = metric;
-    metadata.metrics[metric.id] = { ...rest, collection: null };
-  }
-  return metadata;
-}
-
-export function setupDefinition(
-  metadata: Metadata,
-  metricId: number,
-): LibMetric.MetricDefinition {
-  const provider = LibMetric.metadataProvider(metadata);
-  const metricMeta = LibMetric.metricMetadata(provider, metricId);
-  if (!metricMeta) {
-    throw new Error(`Metric ${metricId} not found`);
-  }
-  return LibMetric.fromMetricMetadata(provider, metricMeta);
-}
+export { createMetricMetadata, setupDefinition };
 
 export function setupDefinitionWithBreakout(
   metadata: Metadata,
